Delete PinnableBufferCache (dotnet/corert#5950)
Port https://github.com/dotnet/coreclr/pull/18360 to CoreRT
